news add article functionality works
