+++ src/scripts/main.js
@@ -2,8 +2,9 @@ import EventListeners from "./tasks/eventHandler"
 import formElements from "./auth/domHandler"
 // import loginRegClick from "./auth/eventHandler"
 import newsFormElements from "./news/newsDomHandler.js"
-import newsApi from "./news/apiHandler.js"
+import newsApi from "./news/newsApiHandler.js"
 import newsRenderDom from "./news/newsEntryDom.js"
+import registerEventListener from "./news/newsEventListeners"
 
 
 import activateEvents from "./events/eventFunctionHandler"
@@ -16,6 +17,7 @@ import API from "./messages/apiHandler.js"
 
 
 import taskForm from "./tasks/domHandler.js"
+
 
 
 sessionStorage.setItem("activeUser", 1)
@@ -77,5 +79,8 @@ newsApi.getArticleEntries()
   .then(response => newsRenderDom.renderArticleEntries(response))
 
 // newsRenderDom.renderArticleEntries()
+registerEventListener()
+   
+
